Refactor zoom
